Harmonik_ortalama güncellendi

+++ Harmonik_ortalama.java
@@ -1,7 +1,7 @@
 package patica;
 public class Harmonik_ortalama {
     public static void main(String[] args) {
-        int[] num={1, 2, 3, 4, 5};
+        double[] num={1, 2, 3, 4, 5};
         double harmonik_seri=0.0;
         for (int i = 0; i < num.length; i++) {
             harmonik_seri+=1/(num[i]);
